Add user dtos and validations

+++ src/controller/user.controller.ts
@@ -2,9 +2,13 @@ import { NextFunction, Request, Response, Router } from 'express';
 import { Service } from 'typedi';
 import NotFoundException from '../exception/not-found.exception';
 import Controller from '../interfaces/controller.interface';
+import requestValidationMiddleware from '../middleware/request-validation.middleware';
 import CollectionService from '../service/collection.service';
 import PostService from '../service/post.service';
 import UserService from '../service/user.service';
+import CreateUserDTO from './dto/create-user.dto';
+import UpdateUserDTO from './dto/update-user.dto';
+import UpgradeUserDTO from './dto/upgrade-user.dto';
 
 @Service()
 export default class UserController implements Controller {
@@ -22,14 +26,15 @@ export default class UserController implements Controller {
         this.router.get(`${this.path}/:id`, this.getUser);
         this.router.get(`${this.path}/:id/posts`, this.getUserPosts);
         this.router.get(`${this.path}/:id/collections`, this.getUserCollections);
-        this.router.post(this.path, this.createUser);
-        this.router.put(`${this.path}/:id`, this.updateUser);
+        this.router.post(this.path, requestValidationMiddleware(CreateUserDTO), this.createUser);
+        this.router.put(`${this.path}/:id/upgrade`, requestValidationMiddleware(UpgradeUserDTO), this.upgradeUser);
+        this.router.put(`${this.path}/:id`, requestValidationMiddleware(UpdateUserDTO), this.updateUser);
         this.router.delete(`${this.path}/:id`, this.deleteUser);
     }
 
     private getUser = async (req: Request, res: Response, next: NextFunction) => {
         const id = req.params.id;
-        const user = await this.userService.getUser(id);
+        const user = await this.userService.getUserById(id);
         if (user) {
             res.send(user);
         } else {
@@ -39,23 +44,50 @@ export default class UserController implements Controller {
 
     private createUser = async (req: Request, res: Response) => {
         res.status(201)
-            .send(await this.userService.createGhostUser(req.body['device_id'], req.body['device_type']));
+            .send(await this.userService.createGhostUser(req.body['deviceId'], req.body['deviceType']));
     }
 
-    private updateUser = async (req: Request, res: Response) => {
+    private upgradeUser = async (req: Request, res: Response, next: NextFunction) => {
         const id = req.params.id;
-        const email = req.body['email'];
-        const password = req.body['password'];
-        const displayName = req.body['display_name'];
-        if (email && password) {
-            res.send(await this.userService.upgradeGhostUser(id, email, password, displayName));
-        } else {
-            res.send(await this.userService.updateUser(id, displayName));
+        const userById = await this.userService.getUserById(id);
+        if (!userById) {
+            next(new NotFoundException('User', id));
         }
+
+        const email = req.body['email'];
+        const userByEmail = await this.userService.getUserByEmail(email);
+        if (userByEmail) {
+            res.status(409).json({ 'message': 'Email already exists' });
+            return;
+        }
+
+        const password = req.body['password'];
+        let displayName = req.body['displayName'];
+        if (!displayName) {
+            displayName = email.split('@', 1)[0];
+        }
+        res.send(await this.userService.upgradeGhostUser(id, email, password, displayName));
     }
 
-    private deleteUser = async (req: Request, res: Response) => {
+    private updateUser = async (req: Request, res: Response, next: NextFunction) => {
         const id = req.params.id;
+        const user = await this.userService.getUserById(id);
+        if (!user) {
+            next(new NotFoundException('User', id));
+        }
+
+        const displayName = req.body['displayName'];
+        res.send(await this.userService.updateUser(id, displayName));
+    }
+
+    private deleteUser = async (req: Request, res: Response, next: NextFunction) => {
+        const id = req.params.id;
+        const user = await this.userService.getUserById(id);
+        if (!user) {
+            next(new NotFoundException('User', id));
+            return;
+        }
+
         await this.userService.deleteUser(id);
         res.sendStatus(204);
     }
